Update game.cpp
Added Nick's code and started with updates. If you are having trouble using this, create a new project in VS Code, and then create your own source file and then copy this file and put it in and run it. This way you don't have to worry about pulling and pushing or merging. Also, create your own branch like Nick did and put your file in there.